#15 feat: 카카오 로그인 후 최초회원 검증 및 새 row 작성 구현
